+++ src/Psr/HttpClientException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace alexeevdv\sms\Psr;
+
+use Psr\Http\Client\ClientExceptionInterface;
+
+final class HttpClientException extends \Exception implements ClientExceptionInterface
+{
+}