+++ src/Psr/RequestFactory.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace alexeevdv\sms\Psr;
+
+use GuzzleHttp\Psr7\Request;
+use Psr\Http\Message\RequestFactoryInterface;
+use Psr\Http\Message\RequestInterface;
+
+final class RequestFactory implements RequestFactoryInterface
+{
+    public function createRequest(string $method, $uri): RequestInterface
+    {
+        return new Request($method, $uri);
+    }
+}