+++ src/provider/MessageId.php
@@ -2,19 +2,21 @@
 
 namespace alexeevdv\sms\provider;
 
-final class MessageId implements \alexeevdv\Sms\Contract\MessageId
+use alexeevdv\Sms\Contract;
+
+final class MessageId implements Contract\MessageId
 {
     /**
      * @var string
      */
     private $messageId;
 
-    public function __construct($messageId)
+    public function __construct(string $messageId)
     {
         $this->messageId = $messageId;
     }
 
-    public function __toString()
+    public function __toString(): string
     {
         return $this->messageId;
     }